pyproject.toml
[tool.poetry]
name = "python-payment-service"
version = "0.1.0"
description = ""
authors = ["kofi12 <kofi12@my.yorku.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
uvicorn = "^0.28.0"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

package-mode = false